Bump Terraform provider to v5.19.0

+++ apis/appsync/v1beta1/zz_datasource_types.go
@@ -218,32 +218,32 @@ type DatasourceParameters struct {
 type DeltaSyncConfigInitParameters struct {
 
 	// The number of minutes that an Item is stored in the data source.
-	BaseTableTTL *float64 `json:"baseTableTtl,omitempty" tf:"base_table_ttl,omitempty"`
+	BaseTableTTL *int64 `json:"baseTableTtl,omitempty" tf:"base_table_ttl,omitempty"`
 
 	// The table name.
 	DeltaSyncTableName *string `json:"deltaSyncTableName,omitempty" tf:"delta_sync_table_name,omitempty"`
 
 	// The number of minutes that a Delta Sync log entry is stored in the Delta Sync table.
-	DeltaSyncTableTTL *float64 `json:"deltaSyncTableTtl,omitempty" tf:"delta_sync_table_ttl,omitempty"`
+	DeltaSyncTableTTL *int64 `json:"deltaSyncTableTtl,omitempty" tf:"delta_sync_table_ttl,omitempty"`
 }
 
 type DeltaSyncConfigObservation struct {
 
 	// The number of minutes that an Item is stored in the data source.
-	BaseTableTTL *float64 `json:"baseTableTtl,omitempty" tf:"base_table_ttl,omitempty"`
+	BaseTableTTL *int64 `json:"baseTableTtl,omitempty" tf:"base_table_ttl,omitempty"`
 
 	// The table name.
 	DeltaSyncTableName *string `json:"deltaSyncTableName,omitempty" tf:"delta_sync_table_name,omitempty"`
 
 	// The number of minutes that a Delta Sync log entry is stored in the Delta Sync table.
-	DeltaSyncTableTTL *float64 `json:"deltaSyncTableTtl,omitempty" tf:"delta_sync_table_ttl,omitempty"`
+	DeltaSyncTableTTL *int64 `json:"deltaSyncTableTtl,omitempty" tf:"delta_sync_table_ttl,omitempty"`
 }
 
 type DeltaSyncConfigParameters struct {
 
 	// The number of minutes that an Item is stored in the data source.
 	// +kubebuilder:validation:Optional
-	BaseTableTTL *float64 `json:"baseTableTtl,omitempty" tf:"base_table_ttl,omitempty"`
+	BaseTableTTL *int64 `json:"baseTableTtl,omitempty" tf:"base_table_ttl,omitempty"`
 
 	// The table name.
 	// +kubebuilder:validation:Optional
@@ -251,7 +251,7 @@ type DeltaSyncConfigParameters struct {
 
 	// The number of minutes that a Delta Sync log entry is stored in the Delta Sync table.
 	// +kubebuilder:validation:Optional
-	DeltaSyncTableTTL *float64 `json:"deltaSyncTableTtl,omitempty" tf:"delta_sync_table_ttl,omitempty"`
+	DeltaSyncTableTTL *int64 `json:"deltaSyncTableTtl,omitempty" tf:"delta_sync_table_ttl,omitempty"`
 }
 
 type DynamodbConfigInitParameters struct {
@@ -556,7 +556,7 @@ type DatasourceStatus struct {
 type Datasource struct {
 	metav1.TypeMeta   `json:",inline"`
 	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// +kubebuilder:validation:XValidation:rule="!('*' in self.managementPolicies || 'Create' in self.managementPolicies || 'Update' in self.managementPolicies) || has(self.forProvider.type) || (has(self.initProvider) && has(self.initProvider.type))",message="spec.forProvider.type is a required parameter"
+	// +kubebuilder:validation:XValidation:rule="!('*' in self.managementPolicies || 'Create' in self.managementPolicies || 'Update' in self.managementPolicies) || has(self.forProvider.type) || has(self.initProvider.type)",message="type is a required parameter"
 	Spec   DatasourceSpec   `json:"spec"`
 	Status DatasourceStatus `json:"status,omitempty"`
 }
